anime_finder_bot/requirements.txt
python-dotenv~=0.20.0
aiogram~=2.22.1
SQLAlchemy==1.3.23
gino~=1.0.1
asyncpg~=0.26.0
asyncio~=3.4.3
aiohttp~=3.8.1
beautifulsoup4~=4.11.1
aiofiles~=0.8.0
selenium~=4.4.3
arsenic~=21.8
bs4~=0.0.1